move data transformations to api hook
